refactor: 반환 타입 요구 사항에 맞게 수정

+++ src/main/java/coupon/domain/repository/MemberCouponRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface MemberCouponRepository extends JpaRepository<MemberCoupon, Long> {
 
-    long countByMemberIdAndCouponId(Long memberId, Long couponId);
+    int countByMemberIdAndCouponId(Long memberId, Long couponId);
 
     List<MemberCoupon> findAllByMemberId(Long memberId);
 }
